requirements.prod.txt
Django==4.2.9
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0 